refactor: move all helpers into separate files
